Revert "gallery"
This reverts commit aa4e3adadbe04576dce0a36fb78a39d5eda66c1c.
