+++ invokeai/frontend/web/src/features/gallery/components/ImageGrid/GalleryImage.tsx
@@ -27,7 +27,57 @@ import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { PiImageBold } from 'react-icons/pi';
 import { imagesApi } from 'services/api/endpoints/images';
 import type { ImageDTO } from 'services/api/types';
-import { galleryItemContainerSX } from './galleryItemContainerSX';
+
+const galleryImageContainerSX = {
+  containerType: 'inline-size',
+  w: 'full',
+  h: 'full',
+  '.gallery-image-size-badge': {
+    '@container (max-width: 80px)': {
+      '&': { display: 'none' },
+    },
+  },
+  '&[data-is-dragging=true]': {
+    opacity: 0.3,
+  },
+  userSelect: 'none',
+  webkitUserSelect: 'none',
+  position: 'relative',
+  justifyContent: 'center',
+  alignItems: 'center',
+  aspectRatio: '1/1',
+  '::before': {
+    content: '""',
+    display: 'inline-block',
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    pointerEvents: 'none',
+    borderRadius: 'base',
+  },
+  '&[data-selected=true]::before': {
+    boxShadow:
+      'inset 0px 0px 0px 3px var(--invoke-colors-invokeBlue-500), inset 0px 0px 0px 4px var(--invoke-colors-invokeBlue-800)',
+  },
+  '&[data-selected-for-compare=true]::before': {
+    boxShadow:
+      'inset 0px 0px 0px 3px var(--invoke-colors-invokeGreen-300), inset 0px 0px 0px 4px var(--invoke-colors-invokeGreen-800)',
+  },
+  '&:hover::before': {
+    boxShadow:
+      'inset 0px 0px 0px 1px var(--invoke-colors-invokeBlue-300), inset 0px 0px 0px 2px var(--invoke-colors-invokeBlue-800)',
+  },
+  '&:hover[data-selected=true]::before': {
+    boxShadow:
+      'inset 0px 0px 0px 3px var(--invoke-colors-invokeBlue-400), inset 0px 0px 0px 4px var(--invoke-colors-invokeBlue-800)',
+  },
+  '&:hover[data-selected-for-compare=true]::before': {
+    boxShadow:
+      'inset 0px 0px 0px 3px var(--invoke-colors-invokeGreen-200), inset 0px 0px 0px 4px var(--invoke-colors-invokeGreen-800)',
+  },
+} satisfies SystemStyleObject;
 
 interface Props {
   imageDTO: ImageDTO;
@@ -114,7 +164,6 @@ export const GalleryImage = memo(({ imageDTO }: Props) => {
         getInitialData: () => {
           const selection = selectSelection(store.getState());
           const boardId = selectSelectedBoardId(store.getState());
-
           // When we have multiple images selected, and the dragged image is part of the selection, initiate a
           // multi-image drag.
           if (selection.length > 1 && selection.includes(imageDTO.image_name)) {
@@ -195,7 +244,7 @@ export const GalleryImage = memo(({ imageDTO }: Props) => {
     <>
       <Flex
         ref={ref}
-        sx={galleryItemContainerSX}
+        sx={galleryImageContainerSX}
         data-is-dragging={isDragging}
         data-item-id={imageDTO.image_name}
         role="button"
