--- invokeai/frontend/web/src/features/gallery/components/useScrollableGallery.tsx
@@ -1,47 +1,0 @@
-import { useOverlayScrollbars } from 'overlayscrollbars-react';
-import { type RefObject, useState, useEffect } from 'react';
-
-/**
- * Handles the initialization of the overlay scrollbars for the gallery, returning the ref to the scroller element.
- */
-export const useScrollableGallery = (rootRef: RefObject<HTMLDivElement>) => {
-  const [scroller, scrollerRef] = useState<HTMLElement | null>(null);
-  const [initialize, osInstance] = useOverlayScrollbars({
-    defer: true,
-    events: {
-      initialized(osInstance) {
-        // force overflow styles
-        const { viewport } = osInstance.elements();
-        viewport.style.overflowX = `var(--os-viewport-overflow-x)`;
-        viewport.style.overflowY = `var(--os-viewport-overflow-y)`;
-      },
-    },
-    options: {
-      scrollbars: {
-        visibility: 'auto',
-        autoHide: 'scroll',
-        autoHideDelay: 1300,
-        theme: 'os-theme-dark',
-      },
-    },
-  });
-
-  useEffect(() => {
-    const { current: root } = rootRef;
-
-    if (scroller && root) {
-      initialize({
-        target: root,
-        elements: {
-          viewport: scroller,
-        },
-      });
-    }
-
-    return () => {
-      osInstance()?.destroy();
-    };
-  }, [scroller, initialize, osInstance, rootRef]);
-
-  return scrollerRef;
-};
